requirements.txt
PySide6>=6.5.0
Pillow>=9.0.0
requests>=2.28.0

# Optional: For building executable
pyinstaller>=5.0.0